ref(metrics): Move Bucket type to new module
