+++ relay-metrics/src/lib.rs
@@ -67,10 +67,12 @@
 )]
 
 mod aggregation;
+mod bucket;
 mod protocol;
 mod router;
 mod statsd;
 
 pub use aggregation::*;
+pub use bucket::*;
 pub use protocol::*;
 pub use router::*;
